* 0.9.x:
  Added a events for the settings menu items
  Bringing Sphinx search support up to code: broken out to a plugin, now supports multiple sites on a single server.
  Changed to Evan's event style and added an AuthPlugin superclass
  add geo output to statuses in json, xml, atom, rss in API
  Localisation updates from translatewiki.net (2009-11-10)
  Localisation updates from translatewiki.net
  Update pot
  add lat and long parameters to api/statuses/update
  change credential check to work more like other events
  fixup output of object attributes in db error code
  Performance fix for subscription/subscriber lists based on feedback from ops.
  Adjusting indexes to make favorites query more efficient, based on feedback from ops.
  Revert untested code; spews PHP notice warnings on every page view: "just sent a http 200 for the check-fancy from install.php"
  Added hook for the Group navigation items
  Updated block @title text (shouldn't say from group)
  Updated group block markup
  Revert "Remove more contractions"

+++ classes/Status_network.php
@@ -57,14 +57,16 @@ class Status_network extends DB_DataObject
         $config['db']['ini_'.$dbname] = INSTALLDIR.'/classes/status_network.ini';
         $config['db']['table_status_network'] = $dbname;
 
-        self::$cache = new Memcache();
+        if (class_exists('Memcache')) {
+            self::$cache = new Memcache();
 
-        if (is_array($servers)) {
-            foreach($servers as $server) {
-                self::$cache->addServer($server);
+            if (is_array($servers)) {
+                foreach($servers as $server) {
+                    self::$cache->addServer($server);
+                }
+            } else {
+                self::$cache->addServer($servers);
             }
-        } else {
-            self::$cache->addServer($servers);
         }
 
         self::$base = $dbname;
@@ -76,6 +78,10 @@ class Status_network extends DB_DataObject
 
     static function memGet($k, $v)
     {
+        if (!self::$cache) {
+            return self::staticGet($k, $v);
+        }
+
         $ck = self::cacheKey($k, $v);
 
         $sn = self::$cache->get($ck);
@@ -92,10 +98,12 @@ class Status_network extends DB_DataObject
 
     function decache()
     {
-        $keys = array('nickname', 'hostname', 'pathname');
-        foreach ($keys as $k) {
-            $ck = self::cacheKey($k, $this->$k);
-            self::$cache->delete($ck);
+        if (self::$cache) {
+            $keys = array('nickname', 'hostname', 'pathname');
+            foreach ($keys as $k) {
+                $ck = self::cacheKey($k, $this->$k);
+                self::$cache->delete($ck);
+            }
         }
     }
